+++ src/app/app.module.ts
@@ -95,6 +95,7 @@ import { ProjectRequestsComponent } from './projects/project-requests/project-re
 import { TeamRolesComponent } from './setups/team-roles/team-roles.component';
 import { DisplayProjectComponent } from './projects/display-project/display-project.component';
 import { UnitTestComponent } from './unit-test/unit-test.component';
+import { TransferProjectsComponent } from './setups/transfer-projects/transfer-projects.component';
 import { MatplanComponent } from './mat-plan/matplan/matplan.component';
 
 
@@ -155,6 +156,7 @@ import { MatplanComponent } from './mat-plan/matplan/matplan.component';
     TeamRolesComponent,
     DisplayProjectComponent,
     UnitTestComponent,
+    TransferProjectsComponent,
     MatplanComponent
   ],
   imports: [