+++ src/app/fte-entry/team/fte-entry.component.ts
@@ -72,6 +72,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
   newPlanName: string;
   planNameRegex: any;
   employeeVisible = new Array;
+  projectVisible = new Array;
   displayNewPlanModal: boolean;
   loginAsEmail: string;
   loginAsID: any;
@@ -208,6 +209,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
 
     if (permRes.length > 0) {
       this.loginAsEmail = this.authService.loggedInUser.managerEmailAddress;
+      // this.loginAsEmail = 'ermina_chua@keysight.com';
       const res2 = await this.apiDataEmployeeService.getEmployeeData(this.loginAsEmail).toPromise();
       this.loginAsID = res2[0].EmployeeID;
       this.displayAdminViewMessage = true;
@@ -238,7 +240,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
   onTestFormClick() {
     // console.log('form object (this.form):');
     // console.log(this.FTEFormGroup);
-    console.log('this.FTEFormGroup.value.FTEFormArray', this.FTEFormGroup.value.FTEFormArray);
+    // console.log('this.FTEFormGroup.value.FTEFormArray', this.FTEFormGroup.value.FTEFormArray);
     // console.log('fte-project-visible array');
     // console.log('teamFTE', this.teamFTEs);
     // console.log('allTeamFTE', this.allTeamFTEs);
@@ -253,7 +255,12 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
     // console.log('this.fteMonthsChart', this.fteMonthsChart)
     // console.log('this.fteChartData', this.fteChartData)
     // console.log('this.employeeTotals', this.employeeTotals)
+    // console.log('this.projectVisible', this.projectVisible)
+    // console.log('this.filterProjects', this.filterProjects)
+    // console.log('this.filterProjectsModel', this.filterProjectsModel)
+    // console.log('this.filterProjectsOptions', this.filterProjectsOptions)
     // this.updateEmployeeTotals();
+    // this.updateProjectFilters()
 
   }
 
@@ -281,19 +288,31 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
       this.setCurrentMonthYear()])
     .then(async res => {
         if (this.defaultPlan === undefined) {
-          this.onCreateNewPlanClick(this.teamEditableMembers, this.loginAsEmail, 'Plan 1');
+          await this.onCreateNewPlanClick(this.teamEditableMembers, this.loginAsEmail, 'Plan 1');
+          this.allProjects.forEach(proj => {
+            this.projectVisible.push(true);
+          });
         } else {
           await this.getPlan(this.loginAsEmail, this.defaultPlan);
+          this.updateProjectFilters(true);
           this.comparePlanToFTE();
           this.createFtePlanningChartData(this.allProjects);
           this.checkDisableDeletePlan();
           this.updateEmployeeTotals();
           this.setEmployeeTotalsBorder();
           this.displayFTETable = true;
+          this.allProjects.forEach(proj => {
+            this.projectVisible.push(true);
+          });
         }
         this.teamOrgStructure.forEach(emp => {
           this.employeeVisible.push(true);
         });
+        // if (this.allProjects.length > 0) {
+        //   this.allProjects.forEach(proj => {
+        //     this.projectVisible.push(true);
+        //   });
+        // }
         this.createFtePlanningChartXAxis();
       }
     );
@@ -370,6 +389,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
   }
 
   onModalClosed(selectedProject: any) {
+    // console.log('this.teamFTEs', this.teamFTEs)
     this.display = true;  // make sure FTE entry form is visible
     setTimeout(() => {
       this.showProjectsModal = false;
@@ -385,6 +405,9 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
         currentProjectsList.push(project[0].projectID);
       }
     });
+
+    // add the project to the projectvisible array so it shows as soon as it is added
+    this.projectVisible.push(true);
 
     const alreadyExists = currentProjectsList.find( value => {
       return value === selectedProject.ProjectID;
@@ -408,7 +431,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
 
 
   onFTEChange(i, j, value) {
-    // console.log(`fte entry changed for project ${i}, month ${j}, with value ${value}`);
+    // console.log(`fte entry changed for project ${i}, employee ${j}, with value ${value}`);
     this.displayNewPlanModal = false;
 
     value = Number(value);
@@ -612,13 +635,17 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
         res => {
           this.cacheService.raiseToast('success', res.message);
 
+          Promise.all([
           // rebuild the FTE entry page
-          this.getPlan(this.loginAsEmail, this.currentPlan);
-          this.buildFteEntryForm();
-          this.displayFTETable = true;
-          this.FTEFormGroup.markAsUntouched();
-          this.checkDisableDeletePlan();
-          this.displayNewPlanModal = true;
+          this.getPlan(this.loginAsEmail, this.currentPlan),
+          this.buildFteEntryForm(),
+          this.displayFTETable = true,
+          this.FTEFormGroup.markAsUntouched(),
+          this.checkDisableDeletePlan(),
+          this.displayNewPlanModal = true
+          ]).then(next => {
+            this.updateProjectFilters(false);
+          });
         },
         err => {
           console.log(err);
@@ -708,7 +735,6 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
   }
 
   buildFteEntryForm() {
-
     this.FTEFormGroup = this.fb.group({
       FTEFormArray: this.fb.array([])
     });
@@ -724,7 +750,6 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
   }
 
   addProjectToFormArray(FTEFormArray: FormArray, proj: TeamFTEs, newProject: boolean) {
-
     const projFormArray = this.fb.array([]); // instantiating a temp formarray for each project
 
     this.teamOrgStructure.forEach(employee => {
@@ -769,7 +794,6 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
   }
 
   addNewProjectToForm(projectID: number, projectName: string) {
-
     const newProject = new TeamFTEs;
     newProject.userID = this.authService.loggedInUser.id;
     newProject.projectID = projectID;
@@ -788,12 +812,16 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
      this.FTEFormGroupLive.push(this.FTEFormGroup.value.FTEFormArray[this.FTEFormGroup.value.FTEFormArray.length - 1]);
      this.FTEFormGroupLive.push(tempArr);
 
-     this.teamFTEs.push({
-       planName: this.currentPlan,
-       projectID: newProject.projectID,
-       projectName: newProject.projectName,
-       allocations: newProject.allocations
-     });
+    // CONFUSED!!!!!
+    // For some reason, when onModalClosed is activated and this chunk of code below is uncommented,
+    //  this.teamFTEs adds the new project TWICE. Cannot find where teamFTEs is being used when user click the "select"
+    //  button on the projectsModal.
+    //  this.teamFTEs.push({
+    //    planName: this.currentPlan,
+    //    projectID: newProject.projectID,
+    //    projectName: newProject.projectName,
+    //    allocations: newProject.allocations
+    //  });
 
      this.allTeamFTEs.push({
        planName: this.currentPlan,
@@ -802,7 +830,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
        allocations: newProject.allocations
      });
 
-     this.allProjects.push(FTEFormArray.controls[FTEFormArray.controls.length - 1]);
+    //  this.allProjects.push(FTEFormArray.controls[FTEFormArray.controls.length - 1]);
      this.projects = FTEFormArray.controls;
 
      // Update the project filter to add the newly added project
@@ -959,8 +987,8 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
   async getTeam(email: string): Promise<any> {
     this.teamOrgStructure = await this.apiDataOrgService.getEmployeeList(email).toPromise();
     this.cacheService.teamEmployeeList = this.teamOrgStructure;
-    this.displayEmployees = this.teamOrgStructure;
-    this.filteredEmployees = this.displayEmployees;
+    // this.displayEmployees = this.teamOrgStructure;
+    this.filteredEmployees = this.teamOrgStructure;
     this.updateEmployeeFilters();
     this.buildTeamEditableMembers();
   }
@@ -1006,7 +1034,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
     this.updateEmployeeTotals();
     this.setEmployeeTotalsBorder();
 
-    this.updateProjectFilters();
+    this.updateProjectFilters(true);
 
     this.planList.splice(0, 0, {planName: planName});
     this.defaultPlan = this.planList[0].planName;
@@ -1022,11 +1050,6 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
     this.teamFTEsFlat = res.flat;
     this.teamFTEsFlatLive = this.teamFTEsFlat;
 
-    // Get launch date from first element
-    if (this.allTeamFTEs[0].allocations[0].launchDate) {
-      this.launchDate = moment(this.allTeamFTEs[0].allocations[0].launchDate).format('YYYY-MM-DD');
-    }
-
     this.buildFteEntryForm(); // initialize the FTE Entry form, which is dependent on FTE data being retrieved
     this.display = true;  // ghetto way to force rendering after FTE data is fetched
     const FTEFormArray = <FormArray>this.FTEFormGroup.controls.FTEFormArray;
@@ -1040,7 +1063,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
     this.updateEmployeeTotals();
     this.setEmployeeTotalsBorder();
 
-    this.updateProjectFilters();
+    // this.updateProjectFilters();
   }
 
   async comparePlanToFTE() {
@@ -1114,7 +1137,9 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
     this.filterEmployeesDropdown.renderFilteredOptions = this.filterEmployeesOptions;
   }
 
-  updateProjectFilters() {
+  updateProjectFilters(init: boolean) {
+    const tempFilterList = this.filterProjects;
+
     // Update list for project dropdown filter
     this.filterProjectsModelStaging = [];
     this.filterProjectsOptionsStaging = [];
@@ -1129,6 +1154,19 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
 
     this.filterProjectsOptions = this.filterProjectsOptionsStaging;
     this.filterProjectsModel = this.filterProjectsModelStaging;
+
+    // // Set the filterProjectsModel back to the selections user had originally
+    // if (!init) {
+    //   this.filterProjectsModel = [];
+    //   for (let i = 0; i < this.filterProjectsOptions.length; i++) {
+    //     for (let j = 0; j < tempFilterList.length; j++) {
+    //       if (this.filterProjectsOptions[i].name === tempFilterList[j].name) {
+    //         this.filterProjectsModel.push(this.filterProjectsOptions[i].id);
+    //       }
+    //     }
+    //   }
+    // }
+
     // For some reason, this does not render dynamically, so this will force the rendering
     //  by accessing the model and updating the renderFilteredOptions manually
     this.filterProjectsDropdown.renderFilteredOptions = this.filterProjectsOptions;
@@ -1314,7 +1352,6 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
   // Angular-2-MultiSelect-Dropdown functions
 
   onFilterEmployeesChange() {
-
     this.filterEmployees = [];
 
     // create an array of all items checked in the filter box with their employees since filterEmployeesModel is only an array of indexes
@@ -1330,6 +1367,7 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
     // teamOrgStructure will always contain the full employee list
     this.filteredEmployees = this.teamOrgStructure.filter(o1 => this.filterEmployees.some(o2 => o1.fullName === o2.name));
 
+    // update the employee visible array. first set all to false, then selected employees to true
     for (let i = 0; i < this.employeeVisible.length; i++) {
       this.employeeVisible[i] = false;
     }
@@ -1337,8 +1375,8 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
       this.employeeVisible[this.filterEmployees[j].id - 1] = true;
     }
 
-    this.buildFteEntryForm(); // initialize the FTE Entry form, which is dependent on FTE data being retrieved
-    this.display = true;  // ghetto way to force rendering after FTE data is fetched
+    // this.buildFteEntryForm(); // initialize the FTE Entry form, which is dependent on FTE data being retrieved
+    // this.display = true;  // ghetto way to force rendering after FTE data is fetched
   }
 
   onFilterProjectsChange() {
@@ -1356,12 +1394,17 @@ export class FteEntryTeamComponent implements OnInit, OnDestroy, ComponentCanDea
     // update the projects list with the filtered list
     // allProjects will always contain the full project list
     this.projects = this.allProjects.filter(o1 => this.filterProjects.some(o2 => o1.projectName === o2.name));
-    // update the teamFTEs to remove the projects from the display
-    // allTeamFTEs will always contain the full teamFTE list
-    this.teamFTEs = this.allTeamFTEs.filter(o1 => this.projects.some(o2 => o1.projectName === o2.projectName));
 
-    this.buildFteEntryForm(); // initialize the FTE Entry form, which is dependent on FTE data being retrieved
-    this.display = true;  // ghetto way to force rendering after FTE data is fetched
+    // update the project visible array. first set all to false, then selected projects to true
+    for (let i = 0; i < this.projectVisible.length; i++) {
+      this.projectVisible[i] = false;
+    }
+    for (let j = 0; j < this.filterProjects.length; j++) {
+      this.projectVisible[this.filterProjects[j].id - 1] = true;
+    }
+
+    // this.buildFteEntryForm(); // initialize the FTE Entry form, which is dependent on FTE data being retrieved
+    // this.display = true;  // ghetto way to force rendering after FTE data is fetched
 
     this.createFtePlanningChartData(this.projects);
   }